Re-architect store implementation tools to group by concern, implement bump.
