+++ lib/mnemonix/store.ex
@@ -7,11 +7,11 @@ defmodule Mnemonix.Store do
 
       iex> Mnemonix.Store.start_link(Mnemonix.Map.Store, name: Store)
       iex> Mnemonix.put(Store, :foo, "bar")
-      iex> Mnemonix.get(Store, :foo)
-      "bar"
+      iex> Mnemonix.fetch(Store, :foo)
+      {:ok, "bar"}
       iex> Mnemonix.delete(Store, :foo)
-      iex> Mnemonix.get(Store, :foo)
-      nil
+      iex> Mnemonix.fetch(Store, :foo)
+      :error
   """
 
   @typedoc """
@@ -120,27 +120,7 @@ defmodule Mnemonix.Store do
 
   use GenServer
 
-
-  @doc false
-
-  @spec init({adapter, opts}) ::
-    {:ok, state} |
-    {:ok, state, timeout | :hibernate} |
-    :ignore |
-    {:stop, reason} when reason: term, timeout: pos_integer
-
-  def init({adapter, opts}) do
-    case adapter.init(opts) do
-      {:ok, state}          -> {:ok, new(adapter, opts, state)}
-      {:ok, state, timeout} -> {:ok, new(adapter, opts, state), timeout}
-      other                 -> other
-    end
-  end
-
-  defp new(adapter, opts, state) do
-    %__MODULE__{adapter: adapter, opts: opts, state: state}
-  end
-
+  use Mnemonix.Store.Lifecycle.Callbacks
 
   @doc false
 
@@ -157,151 +137,7 @@ defmodule Mnemonix.Store do
       reason: term,
       timeout: pos_integer
 
-  ####
-  # CORE
-  ##
-
-  def handle_call({:delete, key}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.delete(store, key) do
-      {:ok, store}         -> {:reply, :ok, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  # TODO: expiry
-  # def handle_call({:expires, key, time}, _, %__MODULE__{adapter: adapter}) do
-  #   case adapter.expires(store, key, time) do
-  #     {:ok, store}         -> {:reply, :ok, store}
-  #     {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-  #   end
-  # end
-
-  def handle_call({:fetch, key}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.fetch(store, key) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:put, key, value}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.put(store, key, value) do
-      {:ok, store}         -> {:reply, :ok, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  ####
-  # MAP FUNCTIONS
-  ##
-
-  def handle_call({:fetch!, key}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.fetch!(store, key) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:get, key}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.get(store, key) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:get, key, default}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.get(store, key, default) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:get_and_update, key, fun}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.get_and_update(store, key, fun) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:get_and_update!, key, fun}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.get_and_update!(store, key, fun) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:get_lazy, key, fun}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.get_lazy(store, key, fun) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:has_key?, key}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.has_key?(store, key) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:pop, key}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.pop(store, key) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:pop, key, default}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.pop(store, key, default) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:pop_lazy, key, fun}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.pop_lazy(store, key, fun) do
-      {:ok, store, value}  -> {:reply, {:ok, value}, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:put_new, key, value}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.put_new(store, key, value) do
-      {:ok, store}         -> {:reply, :ok, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:put_new_lazy, key, fun}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.put_new_lazy(store, key, fun) do
-      {:ok, store}         -> {:reply, :ok, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:update, key, initial, fun}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.update(store, key, initial, fun) do
-      {:ok, store}         -> {:reply, :ok, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-  def handle_call({:update!, key, fun}, _, store = %__MODULE__{adapter: adapter}) do
-    case adapter.update!(store, key, fun) do
-      {:ok, store}         -> {:reply, :ok, store}
-      {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-    end
-  end
-
-
-  @doc false
-
-  @spec terminate(reason, t) :: reason
-    when reason: :normal | :shutdown | {:shutdown, term} | term
-
-  def terminate(reason, store = %__MODULE__{adapter: adapter}) do
-    with {:ok, reason} <- adapter.teardown(reason, store) do
-      reason
-    end
-  end
+  use Mnemonix.Store.Map.Callbacks
+  use Mnemonix.Store.Bump.Callbacks
 
 end
